fix(properties): fix regressions in `status` and `pipestatus` handling
`--status` can be an empty string in some cases with zsh, and it looks like this can't be ignored with clap-derive right now, as a fix switch to `String` which permits empty values.
clap v3 also changed how `Vec<_>` arguments are treated. `multiple_values` can restore the old behavior, but this breaks when passing an empty string. To work around this, quote the arguments in the init scripts and set `value_delimiter = ' '`.