chore: add `bench` profile for faster release builds.
The bench profile is configured like an ordinary release, as release
is configured to optimize for a published production binary. Doing so,
however, is very slow and for testing release builds, it's much faster
to have an escape hatch. The `bench` is the way to do it.

It's 82s vs 20s or better.